Merging recode with uncertainty branch
Adds capability for recoding a station to a location which is related
to the original location by a colocation constraint (horizontal and
vertical uncertainty).

Addresses issue #62, but with some suggested features not yet implemented.

Fixes issue #63 (segfault when using recoding with point data)

Modifies station csv output file to include floating station statistics.

Tidies date formats in output listings to remove time component where it is 00:00:00.