feat: Rationalise choice of orthometric/ellipsoidal heights
Fixes #111.  This provides a simpler an more intuitive handling
of ellipsoidal and orthometric heights.  A coordinate file can contain
one or the other.  This can be defined explicitly.  If it is not then
it now is based on the coordinate system - if it includes a vertical
datum eg NZGD2000/NZVD2016 then the default is orthometric heights,
otherwise it is ellipsoidal.

BREAKING CHANGE: This may change the interpretation of a coordinate
file that does not explicitly state the type of height coordinates.